requirements.txt

PyGithub
python-dateutil
python-dotenv
requests
openai
notion-client
slack-sdk
